+++ src/main/kotlin/com/github/sam0delkin/intellijpsa/psi/PsiElementModelHelper.kt
@@ -26,8 +26,48 @@ class PsiElementModelHelper {
                 return false
             }
 
+            if (null !== pattern.anyParent && null === model.parent) {
+                return false
+            }
+
+            if (null !== pattern.anyParent) {
+                var currentElement = model.parent
+
+                while (true) {
+                    if (null === currentElement) {
+                        return false
+                    }
+
+                    if (matches(currentElement, pattern.anyParent!!)) {
+                        return true
+                    }
+
+                    currentElement = currentElement.parent
+                }
+            }
+
             if (null !== pattern.prev && null === model.prev) {
                 return false
+            }
+
+            if (null !== pattern.anyPrev && null === model.prev) {
+                return false
+            }
+
+            if (null !== pattern.anyPrev) {
+                var currentElement = model.prev
+
+                while (true) {
+                    if (null === currentElement) {
+                        return false
+                    }
+
+                    if (matches(currentElement, pattern.anyPrev!!)) {
+                        return true
+                    }
+
+                    currentElement = currentElement.prev
+                }
             }
 
             if (null !== pattern.prev && !matches(model.prev!!, pattern.prev!!)) {
@@ -36,6 +76,26 @@ class PsiElementModelHelper {
 
             if (null !== pattern.next && null === model.next) {
                 return false
+            }
+
+            if (null !== pattern.anyNext && null === model.next) {
+                return false
+            }
+
+            if (null !== pattern.anyNext) {
+                var currentElement = model.next
+
+                while (true) {
+                    if (null === currentElement) {
+                        return false
+                    }
+
+                    if (matches(currentElement, pattern.anyNext!!)) {
+                        return true
+                    }
+
+                    currentElement = currentElement.next
+                }
             }
 
             if (null !== pattern.next && !matches(model.next!!, pattern.next!!)) {
@@ -60,8 +120,11 @@ class PsiElementModelHelper {
                 withType = model.elementType,
                 withText = model.text,
                 parent = if (model.parent != null) toPattern(model.parent!!) else null,
+                anyParent = null,
                 prev = if (model.prev != null) toPattern(model.prev!!) else null,
+                anyPrev = null,
                 next = if (model.next != null) toPattern(model.next!!) else null,
+                anyNext = null,
                 withOptions = PropertyAccessor.convertToPropertyPathMap(model.options, "options"),
             )
     }